refactor: move golang file to go dir (#7)
* refactor: move golang file to go dir

* ci: add golang and commit lint ci yaml

---------
