+++ .github/workflows/commit_lint.yml
@@ -1,0 +1,49 @@
+name: Lint Commit Messages
+
+on:
+  push:
+    branches:
+      - master
+      - develop
+
+  pull_request:
+    branches:
+      - master
+      - develop
+
+jobs:
+  commit-lint:
+    strategy:
+      matrix:
+        node-version: [14.x]
+        os: [ ubuntu-20.04 ]
+    runs-on: ${{ matrix.os }}
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+
+      - uses: actions/setup-node@v3
+        with:
+          node-version: ${{ matrix.node-version }}
+
+      - uses: actions/cache@v3
+        with:
+          path: |
+            ~/.npm
+            **/node_modules
+          key: ${{ runner.os }}-node-${{ hashFiles('**/package-lock.json') }}
+          restore-keys: |
+            ${{ runner.os }}-node-
+
+      - name: Install Deps
+        run: |
+          npm install -g commitlint-plugin-function-rules @commitlint/cli
+          npm install --save-dev commitlint-plugin-function-rules @commitlint/cli
+
+      - uses: wagoid/commitlint-github-action@v5
+        id: commitlint
+        env:
+          NODE_PATH: ${{ github.workspace }}/node_modules
+        with:
+          configFile: /github/workspace/.github/commitlint.config.js
