+++ go/Makefile
@@ -1,0 +1,6 @@
+SHELL := /bin/bash
+
+.PHONY: all test
+
+test:
+	go test ./...
