+++ go/hash/checksum.go
