+++ go/hash/hash.go
@@ -8,11 +8,11 @@ import (
 
 	"github.com/rs/zerolog/log"
 
-	"github.com/bnb-chain/greenfield-common/redundancy"
+	"github.com/bnb-chain/greenfield-common/go/redundancy"
 )
 
-// ComputerHash split the reader into segment, ec encode the data, compute the hash roots of pieces,
-// and return the hash result array list and data size
+// ComputerHash split the reader into segment, ec encode the data, compute the hash roots of pieces
+// return the hash result array list and data size
 func ComputerHash(reader io.Reader, segmentSize int64, dataShards, parityShards int) ([]string, int64, error) {
 	var segChecksumList [][]byte
 	var result []string
@@ -60,7 +60,7 @@ func ComputerHash(reader io.Reader, segmentSize int64, dataShards, parityShards 
 	result = append(result, segmentRootHash)
 
 	// compute the hash root of pieces of the SecondarySP
-	var wg = &sync.WaitGroup{}
+	wg := &sync.WaitGroup{}
 	spLen := len(encodeData)
 	wg.Add(spLen)
 	hashList := make([]string, spLen)
