+++ go/hash/hash_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/bnb-chain/greenfield-common/redundancy"
+	"github.com/bnb-chain/greenfield-common/go/redundancy"
 )
 
 const (
